--- src/backend/src/main/java/RunningMachines/R2R/domain/course/entity/Difficulty.java
@@ -1,5 +1,0 @@
-package RunningMachines.R2R.domain.course.entity;
-
-public enum Difficulty {
-    LOW, MEDIUM, HIGH
-}
