+++ src/backend/src/main/java/RunningMachines/R2R/domain/running/dto/GpxResponseDto.java
@@ -1,4 +1,4 @@
-package RunningMachines.R2R.domain.course.dto;
+package RunningMachines.R2R.domain.running.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,9 +7,7 @@ import java.util.List;
 
 @Getter
 @AllArgsConstructor
-public class CourseResponseDto {
+public class GpxResponseDto {
     private String fileName;
     private List<String> waypoints;
-    private double distance;
-    private List<String> tags;
 }
