+++ src/backend/src/main/java/RunningMachines/R2R/domain/running/entity/Difficulty.java
@@ -1,0 +1,5 @@
+package RunningMachines.R2R.domain.running.entity;
+
+public enum Difficulty {
+    LOW, MEDIUM, HIGH
+}
