+++ src/backend/src/main/java/RunningMachines/R2R/domain/user/entity/User.java
@@ -3,9 +3,9 @@ package RunningMachines.R2R.domain.user.entity;
 import RunningMachines.R2R.domain.community.entity.Comment;
 import RunningMachines.R2R.domain.community.entity.CommentLike;
 import RunningMachines.R2R.domain.community.entity.Post;
-import RunningMachines.R2R.domain.course.entity.CourseLike;
-import RunningMachines.R2R.domain.course.entity.Review;
-import RunningMachines.R2R.domain.course.entity.UserCourse;
+import RunningMachines.R2R.domain.running.entity.CourseLike;
+import RunningMachines.R2R.domain.running.entity.Review;
+import RunningMachines.R2R.domain.running.entity.UserCourse;
 import RunningMachines.R2R.domain.crew.entity.CrewPost;
 import RunningMachines.R2R.domain.crew.entity.CrewPostComment;
 import RunningMachines.R2R.domain.crew.entity.CrewUser;
